Fix typeahead.js
Typeahead.js functionality was broken due to removal of the 3rd parameter for 'typeahead:selected' callback. This is not going to be fixed due to typeahead.js now being as abandoned as bootstrap-tagsinput.

Fortunately, the version of this script in dist/ had some updated code (as part of the same commit and with no comments, SMH) that works. I have simply replaced the broken block with the working one.